리스트 요소 레이아웃 작업
