+++ lib/ui/home/home_page.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:project_name_change/ui/home/widgets/product.dart';
+import 'package:project_name_change/app/constants/app_constants.dart';
+import 'package:project_name_change/ui/home/widgets/product_for_list.dart';
 
 class HomePage extends StatelessWidget {
   const HomePage({super.key});
@@ -10,7 +11,7 @@ class HomePage extends StatelessWidget {
       body: SafeArea(
         child: Column(
           children: [
-            Product(),
+            ProductForList(AppConstants.baseProductList[0]),
           ]
         ),
       ),
